+++ src/pages/SignupPage.tsx
@@ -132,6 +132,7 @@ export default function SignupPage() {
             id="email"
             name="email"
             placeholder="이메일을 입력해 주세요."
+            autoFocus
             onChange={handleChangeUser}
           />
           <ErrorMessage>{isError.email.error && isError.email.message}</ErrorMessage>
